cli/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argcomplete==3.1.4
Authlib==1.6.8
black==26.1.0
blinker==1.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
conan==2.25.2
coverage==7.13.4
cryptography==41.0.7
dbus-python==1.3.2
distro==1.9.0
dparse==0.6.4
fasteners==0.20
filelock==3.24.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
librt==0.8.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.2
mdurl==0.1.2
mypy==1.19.1
mypy_extensions==1.1.0
nltk==3.9.2
oauthlib==3.2.2
packaging==24.0
patch-ng==1.18.1
pathspec==1.0.4
platformdirs==4.9.2
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyGObject==3.48.2
PyJWT==2.7.0
pyparsing==3.1.1
pytest==9.0.2
pytest-cov==7.0.0
python-apt==2.7.7+ubuntu5.2
python-dateutil==2.9.0.post0
pytokens==0.4.1
PyYAML==6.0.1
rdflib==7.6.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
ruamel.yaml==0.19.1
ruff==0.15.2
safety==3.7.0
safety-schemas==0.0.16
shellingham==1.5.4
six==1.17.0
tenacity==9.1.4
toml==0.10.2
tomlkit==0.14.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wadllib==1.3.6
xmltodict==0.13.0
-e git+http://local_proxy@127.0.0.1:33482/git/seanchatmangpt/yawl@683bb399aaf6d146d20d9ead46ad28a6bdd53f94#egg=yawl_cli&subdirectory=cli
yq==3.1.0
